Created and segregated metaThemeUpdater
